refactor: sionna 네이밍 정리 및 ai_runtime 인프라로 통합
