+++ apps/ai_api/app/api/deps.py
@@ -4,7 +4,7 @@ from app.infrastructure.ai_runtime.inference_runner import (
     run_unet_with_runtime,
     run_yolo_with_runtime,
 )
-from app.infrastructure.sionna.sionna_preview_runner import run_sionna_preview_with_runtime
+from app.infrastructure.ai_runtime.sionna_gateway import run_sionna_with_runtime
 
 
 def get_unet_runner():
@@ -15,5 +15,5 @@ def get_yolo_runner():
     return run_yolo_with_runtime
 
 
-def get_sionna_preview_runner():
-    return run_sionna_preview_with_runtime
+def get_sionna_runner():
+    return run_sionna_with_runtime
