+++ apps/ai_api/app/api/routes/sionna.py
@@ -4,14 +4,14 @@ from __future__ import annotations
 
 from fastapi import APIRouter, Depends
 
-from app.api.deps import get_sionna_preview_runner
+from app.api.deps import get_sionna_runner
 from app.api.errors import AppError, to_http_exception
 from app.presentation.requests.sionna_request_dto import SionnaRunRequestDto
 from app.presentation.responses.sionna_response_dto import (
     SionnaRunResponseDto,
     to_sionna_response,
 )
-from app.usecases.run_sionna_preview_usecase import run_sionna_preview_usecase
+from app.usecases.run_sionna_usecase import run_sionna_usecase
 
 router = APIRouter()
 
@@ -19,12 +19,12 @@ router = APIRouter()
 @router.post("/sionna/run", response_model=SionnaRunResponseDto)
 def post_internal_sionna_run(
     body: SionnaRunRequestDto,
-    runner=Depends(get_sionna_preview_runner),
+    runner=Depends(get_sionna_runner),
 ) -> SionnaRunResponseDto:
     try:
         if body.engine != "sionna_rt":
             raise AppError(status_code=400, detail="only engine=sionna_rt is supported")
-        result = run_sionna_preview_usecase(body, runner)
+        result = run_sionna_usecase(body, runner)
         if result.get("status") == "failed":
             raise AppError(status_code=500, detail=result.get("error") or "Sionna RT run failed")
     except Exception as exc:
