+++ apps/ai_api/app/domain/entities/sionna_entity.py
