+++ apps/ai_api/app/infrastructure/ai_runtime/__init__.py
@@ -1,3 +1,11 @@
+from app.infrastructure.ai_runtime.inference_runner import run_unet_with_runtime, run_yolo_with_runtime
+from app.infrastructure.ai_runtime.sionna_gateway import run_sionna_with_runtime
+
+__all__ = [
+    "run_unet_with_runtime",
+    "run_yolo_with_runtime",
+    "run_sionna_with_runtime",
+]
 from app.infrastructure.ai_runtime.inference_runner import (
     run_unet_with_runtime,
     run_yolo_with_runtime,
