--- apps/ai_api/app/infrastructure/sionna/__init__.py
@@ -1,3 +1,0 @@
-from app.infrastructure.sionna.sionna_preview_runner import run_sionna_preview_with_runtime
-
-__all__ = ["run_sionna_preview_with_runtime"]
