+++ apps/ai_api/app/presentation/requests/sionna_request_dto.py
@@ -4,13 +4,13 @@ from typing import Literal
 
 from pydantic import BaseModel
 
-from app.domain.entities.sionna_preview_entity import (
+from app.domain.entities.sionna_entity import (
     InputSionnaPayloadEntity,
 )
 
 
 class SionnaRunRequestDto(BaseModel):
     engine: Literal["sionna_rt"] = "sionna_rt"
-    run_type: str = "preview"
+    run_type: str = "run"
     floor_id: str | None = None
     input: InputSionnaPayloadEntity
