+++ apps/ai_api/app/usecases/__init__.py
@@ -1,9 +1,9 @@
-from app.usecases.run_sionna_preview_usecase import run_sionna_preview_usecase
+from app.usecases.run_sionna_usecase import run_sionna_usecase
 from app.usecases.run_unet_usecase import run_unet_usecase
 from app.usecases.run_yolo_usecase import run_yolo_usecase
 
 __all__ = [
-    "run_sionna_preview_usecase",
+    "run_sionna_usecase",
     "run_unet_usecase",
     "run_yolo_usecase",
 ]
