+++ apps/ai_api/app/usecases/run_sionna_usecase.py
@@ -5,7 +5,7 @@ from typing import Any
 from app.presentation.requests.sionna_request_dto import SionnaRunRequestDto
 
 
-def run_sionna_preview_usecase(body: SionnaRunRequestDto, runner):
+def run_sionna_usecase(body: SionnaRunRequestDto, runner):
     payload: dict[str, Any] = {
         "engine": body.engine,
         "run_type": body.run_type,
